+++ apps/gateway/src/Dawning.Identity.Infra.Data/Repository/OpenIddict/AuthorizationRepository.cs
@@ -119,7 +119,7 @@ namespace Dawning.Identity.Infra.Data.Repository.OpenIddict
             }
 
             AuthorizationEntity entity = model.ToEntity();
-            return await _context.Connection.InsertAsync(entity, _context.Transaction);
+            return (int)await _context.Connection.InsertAsync(entity, _context.Transaction);
         }
 
         /// <summary>
